% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_approximation.r
\name{errorOUGamma}
\alias{errorOUGamma}
\title{errorOUGamma}
\usage{
errorOUGamma(
  t,
  lambda,
  alpha,
  mu,
  tilde.sigma2,
  x0,
  deltax,
  NpointsOUG,
  Nsim,
  scenario
)
}
\arguments{
\item{t}{time at which to evaluate the IAE;}

\item{lambda}{positive constant rate of the underlying Poisson process;}

\item{alpha}{positive exponential decay rate of the shot noise process}

\item{mu}{positive quantity entering into the scale and rate parameters of the Gamma jump amplitude Jn, see Table 1 of Tamborrino and Lansky, Physica D, 418 (2021) 132845 (TL2021).
This corresponds to the limit first infinitesimal moment, eq. (23)}

\item{tilde.sigma2}{positive quantity entering into the scale and rate parameters of the Gamma jump amplitude of Jn, see Table 1 of Tamborrino and Lansky, Physica D, 418 (2021) 132845 (TL2021).
This corresponds to the limit second infinitesimal moment, eq. (24), see also Table 1}

\item{x0}{non-negative starting condition of the processes;}

\item{deltax}{simulation time step;}

\item{NpointsOUG}{number of points to simulate in the OU-Gamma process.}

\item{Nsim}{number of simulation of the shot noise process;}

\item{scenario}{1 if Jn are chi-square distributed, 2 if Gamma}
}
\value{
IAEs of the shot noise process at time t with respect to the OU process and the limiting non-Gaussian OU-Gamma process
}
\description{
Integrate Absolute error (IAE) of the shot noise vs the OU and OU-Gamma  process
with mean x0*exp(-alpha*t) + mu*(1-exp(-alpha*t))/alpha and standard deviation sqrt(tilde.sigma2*(1-exp(-2*alpha*t))/(2*alpha)) at time t.
Scenario 1 for Chi-square distributed jumps Jn, scenario 2 for Gamma distributed jumps Jn
}
