% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_approximation.r
\name{errorOUIG}
\alias{errorOUIG}
\title{errorOUIG}
\usage{
errorOUIG(t, lambda, alpha, mu, tilde.sigma2, x0, deltax, NpointsOUIG, Nsim)
}
\arguments{
\item{t}{time at which to evaluate the IAE;}

\item{lambda}{positive constant rate of the underlying Poisson process;}

\item{alpha}{positive exponential decay rate of the shot noise process}

\item{mu}{positive quantity entering into the mean and shape parameters of the IG jump distribution Jn (Table 1, Tamborrino and Lansky, Physica D, 418 (2021) 132845, TL2021), corresponding to the limiting first infinitesimal moment
eq. (23);}

\item{tilde.sigma2}{positive quantity entering into the shape parameter of the IG Jn, corresponding also to the
limiting second moment of eq. (24);}

\item{x0}{non-negative starting condition of the processes;}

\item{deltax}{simulation time step;}

\item{NpointsOUIG}{number of points to simulate in the OU-IG process.}

\item{Nsim}{number of simulation of the shot noise process;}
}
\value{
IAEs of the shot noise process at time t with respect to the OU process and the limiting non-Gaussian OU-IG process
}
\description{
Integrate Absolute error (IAE) of the shot noise vs the OU and OU-Inverse Gaussian (OU-IG) process
with mean x0*exp(-alpha*t) + mu*(1-exp(-alpha*t))/alpha and standard deviation sqrt(tilde.sigma2*(1-exp(-2*alpha*t))/(2*alpha)) at time t.
}
