% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_approximation.r
\name{errorOUPoisson}
\alias{errorOUPoisson}
\title{errorOUPoisson}
\usage{
errorOUPoisson(t, lambda, alpha, mu, x0, deltax, Nsim, scenario)
}
\arguments{
\item{t}{time at which to evaluate the IAE;}

\item{lambda}{positive constant rate of the underlying Poisson process;}

\item{alpha}{positive exponential decay rate of the shot noise process}

\item{mu}{positive quantity entering into the jump distribution Jn;}

\item{x0}{non-negative starting condition of the processes;}

\item{deltax}{simulation time step;}

\item{Nsim}{number of simulation of the shot noise;}

\item{scenario}{1 if Jn are Bernoulli distributed, 2 if Poisson}
}
\value{
IAEs of the shot noise process at time t with respect to the OU process and the limiting non-Gaussian OU-Poisson
}
\description{
Integrate Absolute error (IAE) of the shot noise vs the OU and the OU-Poisson process
with mean x0*exp(-alpha*t) + mu*(1-exp(-alpha*t))/alpha and standard deviation sqrt(mu*(1-exp(-2*alpha*t))/(2*alpha)) at time t.
Scenario 1 for Bernoulli distributed jumps Jn, scenario 2 for Poisson distributed jumps Jn
}
