% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotnoise.r
\docType{package}
\name{shotnoise}
\alias{shotnoise}
\title{shotnoise}
\description{
Exact simulation of non-negative shot noise processes with non-negative jump distributions, and of non-Gaussian OU processes, in particular OU-Poisson, OU-Gamma and OU-Inverse Gaussian processes.
}
\author{
Massimiliano Tamborrino
}
