% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotnoise_sim.r
\name{shotnoise_JnChiSq}
\alias{shotnoise_JnChiSq}
\title{shotnoise_JnChiSq}
\usage{
shotnoise_JnChiSq(Tmax, Nsim, x0, alpha, mu, lambda)
}
\arguments{
\item{Tmax}{time at which we want to sample from;}

\item{Nsim}{number of simulated values from Xn(Tmax)}

\item{x0}{non-negative starting condition}

\item{alpha}{positive exponential decay rate}

\item{mu}{positive quantity entering into the shape parameter of the Gamma jump amplitude Jn, see point 2 of Section 3.1 of Tamborrino and Lansky, Physica D, 418 (2021) 132845 (TL2021).}

\item{lambda}{positive intensity of the underlying Poisson process}
}
\value{
Nsim values from Xn(Tmax) of the shot noise process with jumps Jn being chi-square distributed
}
\description{
Simulation of Nsim values from the shot noise process Xn(t) with Chi-square distributed jumps Jn at time Tmax
}
